backend/src/city_agent/agent.py
from google.adk.agents import LlmAgent, SequentialAgent
import asyncio
from rag_pipeline.vector import query_retriever
from ai_api_selector import get_agent_model


async def search_data(query: str) -> str:
    # Offload the (potentially) blocking retriever call to a thread so the
    # async event loop isn't blocked when this tool is used inside an async
    # agent/runtime.
    relevant_data = await asyncio.to_thread(query_retriever, query)
    return relevant_data


ai_api = get_agent_model()

# TODO: Refine this prompt
_ORCHESTRATOR_INSTRUCTIONS = """
You are the CityAgent Orchestrator.
Goal: Accurately route user inquiries regarding City of Ottawa Asset Management to the appropriate sub-agent and synthesize the results.
Available sub-agents:
    Reasoner Agent: Capable of semantic search, data retrieval, and logical deduction based on City records.
Process: Before answering, perform the following internal steps:
    1. Check Scope: Is this specifically about the City of Ottawa AND related to asset management/infrastructure?
        If NO: Politely decline.
        If YES: Proceed to step 2.
    2. Formulate Plan: Identify the specific data points needed to answer the user.
    3. Execute: Call the Reasoner Agent with the necessary context.
Constraint Checklist & Confidence Score:
    Do not answer general knowledge questions (e.g., "What is the capital of Canada?") unless they strictly pertain to asset management (e.g., "What is the maintenance budget for the Capital region's water pipes?").
    If the Reasoner Agent returns no results, admit that the information is unavailable rather than making it up.
"""

# TODO: Refine this prompt
_REASONER_INSTRUCTIONS = """
You are CityAgent Reasoner, a specialized sub-agent for the City of Ottawa. Your sole purpose is to retrieve, analyze, and synthesize technical data regarding city assets to fulfill requests from the Orchestrator.
Available Tools
    search_data: Use this to query the vector database. The return output will contain both content (text) and metadata (source details).
Operational Protocol
When you receive a query, follow this step-by-step process:
    Query Optimization: Convert the user's natural language request into technical keywords likely to be found in official city documentation (e.g., change "road repair rules" to "asphalt maintenance standards").
    Execution: Call search_data.
    Extraction & Verification:
        Analyze the text content to ensure it answers the specific question.
Failure Condition
If search_data returns empty results or results that are completely unrelated to the query, you must output exactly: "I could not find any relevant information."
CRITICAL: Locate the filename and last_updated (or date_modified) fields in the metadata of the relevant chunks.
Citation Protocol
You must provide a citation for every piece of information you retrieve. Do not include data if you cannot attribute it to a specific file. Use the following format for your answers:
    Format: [Fact/Answer] (Source: [filename], Last Updated: [date])
    Example: "The budget for the Rideau Canal maintenance is set at $2.5M (Source: 2024_Capital_Budget_Final.pdf, Last Updated: 2023-12-15)."
"""

reasoner_agent = LlmAgent(
    name="CityAgent_Reasoner",
    model=ai_api,
    description=("The orchestrator's reasoning agent."),
    instruction=(_REASONER_INSTRUCTIONS),
    tools=[search_data],
)


root_agent = LlmAgent(
    name="CityAgent_Orchestrator",
    model=ai_api,
    description=("A helpful assistant for user questions."),
    instruction=(_ORCHESTRATOR_INSTRUCTIONS),
    sub_agents=[reasoner_agent],
)

# Used by the Agent Evaluator for testing purposes
agent = root_agent